mailing/sickle-mailing/requirements.txt
firebase_functions
firebase-admin
email-validator
python-dotenv
